// src/stock/constants/paths.enum.ts
export enum Paths {
    MANY = "many",

    STOCK = "stock",
    PARAM_STOCK = ":stockId",
    PARAM_STOCK_ID = "stockId",

    ITEM = "item",
    PARAM_ITEM = ":itemId",
    PARAM_ITEM_ID = "itemId",

    LOT = "lot",
    PARAM_LOT = ":lotId",
    PARAM_LOT_ID = "lotId",

    MOVEMENT = "movement",
    PARAM_MOVEMENT = ":movementId",
    PARAM_MOVEMENT_ID = "movementId",

    TRANSFER = "transfer",
    PARAM_TRANSFER = ":transferId",
    PARAM_TRANSFER_ID = "transferId",

    MOVEMENTS_LINKER = "movement/linker",
    PARAM_MOVEMENTS_LINKER = ":movementLinkerId",
    PARAM_MOVEMENTS_LINKER_ID = "movementLinkerId"
}